Improved the language following the anon-func example.
It was unnecessarily awkward and compressed to read, plus the less than ideal use of `a` and "a" in the same sentence... using markup to try to fix a writing problem. Alternatively, one could change the example code to use something like `x` or `f` instead of `a`.
In any case, whilst an overly conversational style can be draining and actually hide details, sentences still need to breathe. It's for a person to read, not for a person to *parse* - and even it were, the context to be inferred is a little unreasonable.

What is still missing is something mentioning that functions-as-values don't have to be anonymous, which is what could be inferred from this block of text with the example. There should ideally be an example that defines a global function (instance?) for recursive use, say. Even steal the example from the Go Tour?